Update README.md
This project uses the HC-SR04 ultrasonic sensor to measure the distance between the device and any object placed in front of it. The ultrasonic sensor works by emitting a high-frequency sound pulse through its trigger pin. This sound wave travels through the air until it hits an object, and then bounces back to the sensor’s echo pin.

The Raspberry Pi Pico sends a short 10µs pulse to the trigger pin to start the measurement. It then listens on the echo pin to measure the duration of the returned signal. Using the speed of sound (343 m/s or 0.0343 cm/µs), the code calculates the distance with the formula.
Once the distance is calculated, the code checks if it is less than a certain threshold (e.g., 30 cm). If true, the buzzer is activated, giving an audio warning that something is too close. If the distance is safe (more than the threshold), the buzzer is turned off.

This mechanism is particularly useful in:

*Smart walking sticks for visually impaired people

*Automated proximity alarms

*Object detection systems

*Safety barriers

The entire logic runs continuously in a loop, making it a real-time distance monitoring system using embedded hardware and MicroPython.